requirements.txt
websocket-client==1.6.1
pillow==10.0.0
Jinja2==3.1.2
py-cord==2.4.1
python-dotenv==1.0.0
websockets==11.0.3
aiohttp[speedups]==3.8.5
openai==1.7.2
aiosqlite==0.19.0
tiktoken==0.5.2